requirements.txt
# ============================================================
# Eco Pharm Telegram Bot - Python Dependencies
# ============================================================

# Telegram Bot Framework
aiogram>=3.4.0,<4.0.0

# Database - PostgreSQL
asyncpg>=0.29.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Database - SQLite (for migration only)
aiosqlite>=0.19.0,<1.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Scheduler
APScheduler>=3.10.4,<4.0.0

# Timezone
pytz>=2024.1

# Database migration
alembic>=1.13.0,<2.0.0